Fix a subtle error occuring at rationalization
When there were duplicate roles, both of them would be removed. This
happened because the filter method would just check that the index of
a and b were not the same (securing that it wasn't comparing against
itself). But since it would first visit a, it would see that b were
the same and remove a, and then visit b and see that a were the same -
thereby removing b. By checking that the index of b is higher than a,
we will remove a but not b.

The unit test provided by Schlüter does not in fact check that
duplicates are removed - just that lesser scopes are removed, because
they are shadowed by a parent role.